Se añade creaciones ya ctualizaciones de información

+++ CasaRM.WebApp/CasaRM.WebApp.Repositories/Implementations/GenericRepository.cs
@@ -71,7 +71,7 @@ namespace CasaRM.WebApp.Repositories.Implementations
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public async Task<T> GetByIdAsync(long id)
+        public async Task<T> GetByIdAsync(int id)
         {
             return await _context.Set<T>().FindAsync(id);
         }
